refactor(diff): extract loadDiff, convertToSideBySide, formatCommentsAsPrompt to shared/utils/diff/
Move pure utilities and diff types out of DiffView.tsx into their own modules.
Update the two test files that previously imported from DiffView.
