+++ src/components/views/DiffView.tsx
@@ -19,201 +19,10 @@ import {ViewportCalculator} from '../../shared/utils/viewport.js';
 import {computeUnifiedPerFileIndices, computeSideBySidePerFileIndices} from '../../shared/utils/diffLineIndex.js';
 import {calculateDiffViewportRows} from '../../shared/utils/layout.js';
 import {getLanguageFromFileName} from '../../shared/utils/languageMapping.js';
-
-type DiffLine = {
-  type: 'added'|'removed'|'context'|'header';
-  text: string;
-  fileName?: string;
-  headerType?: 'file' | 'hunk';
-  oldLineIndex?: number; // 1-based original (left) line number from diff output
-  newLineIndex?: number; // 1-based current (right) line number from diff output
-};
-
-type SideBySideLine = {
-  left: {type: 'removed'|'context'|'header'|'empty'; text: string; fileName?: string; headerType?: 'file' | 'hunk'; oldLineIndex?: number; newLineIndex?: number} | null;
-  right: {type: 'added'|'context'|'header'|'empty'; text: string; fileName?: string; headerType?: 'file' | 'hunk'; oldLineIndex?: number; newLineIndex?: number} | null;
-  lineIndex: number; // Original line index for comments and navigation
-};
-
-// Map file extensions to language identifiers for syntax highlighting is now in shared util
-
-async function loadDiff(worktreePath: string, diffType: 'full' | 'uncommitted' = 'full', baseCommitHash?: string): Promise<DiffLine[]> {
-  let diff: string | null = null;
-  
-  if (diffType === 'uncommitted') {
-    // Show only uncommitted changes (working directory vs HEAD)
-    diff = await runCommandAsync(['git', '-C', worktreePath, 'diff', '--no-color', '--no-ext-diff', 'HEAD']);
-  } else {
-    // Show full diff against base branch (default behavior)
-    let target = baseCommitHash;
-    if (!target) {
-      target = 'HEAD~1';
-      const base = findBaseBranch(worktreePath, BASE_BRANCH_CANDIDATES);
-      if (base) {
-        const mb = await runCommandAsync(['git', '-C', worktreePath, 'merge-base', 'HEAD', base]);
-        if (mb) target = mb.trim();
-      }
-    }
-    diff = await runCommandAsync(['git', '-C', worktreePath, 'diff', '--no-color', '--no-ext-diff', target]);
-  }
-  
-  // Collect files and their content
-  const fileContents = new Map<string, DiffLine[]>();
-  
-  // Process main diff if it exists
-  if (diff && diff.trim()) {
-    const raw = diff.split('\n');
-    let currentFileName = '';
-    let currentFileLines: DiffLine[] = [];
-    let oldLineCounter = 1;
-    let newLineCounter = 1;
-    
-    for (const line of raw) {
-      if (line.startsWith('diff --git')) {
-        // Save previous file if it exists
-        if (currentFileName && currentFileLines.length > 0) {
-          fileContents.set(currentFileName, currentFileLines);
-        }
-        
-        // Start new file
-        const parts = line.split(' ');
-        const fp = parts[3]?.slice(2) || parts[2]?.slice(2) || '';
-        currentFileName = fp;
-        currentFileLines = [];
-        currentFileLines.push({type: 'header', text: `📁 ${fp}`, fileName: fp, headerType: 'file'});
-        oldLineCounter = 1;
-        newLineCounter = 1;
-      } else if (line.startsWith('@@')) {
-        // Parse original and new starting line numbers from hunk header
-        const m = line.match(/^@@ -([0-9]+)(?:,([0-9]+))? \+([0-9]+)(?:,([0-9]+))? @@/);
-        if (m) {
-          const oldStart = parseInt(m[1] || '1', 10);
-          const newStart = parseInt(m[3] || '1', 10);
-          oldLineCounter = Math.max(1, oldStart);
-          newLineCounter = Math.max(1, newStart);
-        }
-        const ctx = line.replace(/^@@.*@@ ?/, '');
-        if (ctx) currentFileLines.push({type: 'header', text: ` ▼ ${ctx}`, fileName: currentFileName, headerType: 'hunk'});
-      } else if (line.startsWith('+') && !line.startsWith('+++')) {
-        currentFileLines.push({type: 'added', text: line.slice(1), fileName: currentFileName, newLineIndex: newLineCounter});
-        newLineCounter++;
-      } else if (line.startsWith('-') && !line.startsWith('---')) {
-        currentFileLines.push({type: 'removed', text: line.slice(1), fileName: currentFileName, oldLineIndex: oldLineCounter});
-        oldLineCounter++;
-      } else if (line.startsWith(' ')) {
-        currentFileLines.push({type: 'context', text: line.slice(1), fileName: currentFileName, oldLineIndex: oldLineCounter, newLineIndex: newLineCounter});
-        oldLineCounter++;
-        newLineCounter++;
-      } else if (line === '') {
-        currentFileLines.push({type: 'context', text: ' ', fileName: currentFileName}); // Empty line gets a space so cursor is visible
-      }
-    }
-    
-    // Save last file
-    if (currentFileName && currentFileLines.length > 0) {
-      fileContents.set(currentFileName, currentFileLines);
-    }
-  }
-  
-  // Process untracked files
-  const untracked = await runCommandAsync(['git', '-C', worktreePath, 'ls-files', '--others', '--exclude-standard']);
-  if (untracked) {
-    for (const fp of untracked.split('\n').filter(Boolean)) {
-      const fileLines: DiffLine[] = [];
-      fileLines.push({type: 'header', text: `📁 ${fp} (new file)`, fileName: fp, headerType: 'file'});
-      try {
-        const cat = await runCommandAsync(['bash', '-lc', `cd ${JSON.stringify(worktreePath)} && sed -n '1,200p' ${JSON.stringify(fp)}`]);
-        for (const l of (cat || '').split('\n')) {
-          // Keep all lines including empty ones to faithfully render blank lines
-          fileLines.push({type: 'added', text: l, fileName: fp});
-        }
-      } catch {}
-      fileContents.set(fp, fileLines);
-    }
-  }
-  
-  // Sort files alphabetically and combine their content
-  const sortedFiles = Array.from(fileContents.keys()).sort();
-  const lines: DiffLine[] = [];
-  
-  for (const fileName of sortedFiles) {
-    const fileLines = fileContents.get(fileName);
-    if (fileLines) {
-      lines.push(...fileLines);
-    }
-  }
-  
-  return lines;
-}
-
-function convertToSideBySide(unifiedLines: DiffLine[]): SideBySideLine[] {
-  const sideBySideLines: SideBySideLine[] = [];
-  let lineIndex = 0;
-  let i = 0;
-
-  while (i < unifiedLines.length) {
-    const line = unifiedLines[i];
-    
-    if (line.type === 'header') {
-      // Headers appear on both sides
-      sideBySideLines.push({
-        left: {type: 'header', text: line.text, fileName: line.fileName, headerType: line.headerType},
-        right: {type: 'header', text: line.text, fileName: line.fileName, headerType: line.headerType},
-        lineIndex: lineIndex++
-      });
-      i++;
-    } else if (line.type === 'context') {
-      // Context lines appear on both sides
-      sideBySideLines.push({
-        left: {type: 'context', text: line.text, fileName: line.fileName, oldLineIndex: line.oldLineIndex, newLineIndex: line.newLineIndex},
-        right: {type: 'context', text: line.text, fileName: line.fileName, oldLineIndex: line.oldLineIndex, newLineIndex: line.newLineIndex},
-        lineIndex: lineIndex++
-      });
-      i++;
-    } else if (line.type === 'removed') {
-      // Collect all consecutive removed lines
-      const removedLines: DiffLine[] = [];
-      while (i < unifiedLines.length && unifiedLines[i].type === 'removed') {
-        removedLines.push(unifiedLines[i]);
-        i++;
-      }
-      
-      // Collect all consecutive added lines that follow
-      const addedLines: DiffLine[] = [];
-      while (i < unifiedLines.length && unifiedLines[i].type === 'added') {
-        addedLines.push(unifiedLines[i]);
-        i++;
-      }
-      
-      // Pair them up, filling with empty lines as needed
-      const maxLines = Math.max(removedLines.length, addedLines.length);
-      
-      for (let j = 0; j < maxLines; j++) {
-        const removedLine = removedLines[j] || null;
-        const addedLine = addedLines[j] || null;
-        
-        sideBySideLines.push({
-          left: removedLine ? {type: 'removed', text: removedLine.text, fileName: removedLine.fileName, oldLineIndex: removedLine.oldLineIndex} : {type: 'empty', text: '', fileName: line.fileName},
-          right: addedLine ? {type: 'added', text: addedLine.text, fileName: addedLine.fileName, newLineIndex: addedLine.newLineIndex} : {type: 'empty', text: '', fileName: line.fileName},
-          lineIndex: lineIndex++
-        });
-      }
-    } else if (line.type === 'added') {
-      // Added lines without preceding removed lines
-      sideBySideLines.push({
-        left: {type: 'empty', text: '', fileName: line.fileName},
-        right: {type: 'added', text: line.text, fileName: line.fileName, newLineIndex: line.newLineIndex},
-        lineIndex: lineIndex++
-      });
-      i++;
-    } else {
-      i++;
-    }
-  }
-
-  return sideBySideLines;
-}
-
+import {loadDiff} from '../../shared/utils/diff/loadDiff.js';
+import {convertToSideBySide} from '../../shared/utils/diff/convertToSideBySide.js';
+import {formatCommentsAsPrompt} from '../../shared/utils/diff/formatCommentsAsPrompt.js';
+import type {DiffLine, SideBySideLine, ViewMode, WrapMode} from '../../shared/utils/diff/types.js';
 
 type Props = {
   worktreePath: string;
@@ -224,55 +33,6 @@ type Props = {
   // When viewing a workspace child repo, route comments to the top-level workspace session
   workspaceFeature?: string;
 };
-
-type ViewMode = 'unified' | 'sidebyside';
-type WrapMode = 'truncate' | 'wrap';
-
-// Exported utility for formatting comments into a Claude-friendly prompt
-export function formatCommentsAsPrompt(
-  comments: any[],
-  opts?: {workspaceFeature?: string; project?: string; baseCommitHash?: string}
-): string {
-  let prompt = "Please address the following code review comments:\n\n";
-  if (opts?.workspaceFeature && opts?.project) {
-    prompt += `Context: In workspace '${opts.workspaceFeature}', target child directory: ./${opts.project}\n\n`;
-  }
-
-  const commentsByFile: {[key: string]: typeof comments} = {};
-  comments.forEach(comment => {
-    if (!commentsByFile[comment.fileName]) {
-      commentsByFile[comment.fileName] = [];
-    }
-    commentsByFile[comment.fileName].push(comment);
-  });
-
-  Object.entries(commentsByFile).forEach(([fileName, fileComments]) => {
-    const header = opts?.baseCommitHash ? `File: ${fileName}@${opts.baseCommitHash}` : `File: ${fileName}`;
-    prompt += `${header}\n`;
-    fileComments.forEach(comment => {
-      if (comment.lineIndex !== undefined) {
-        // Normal line with line number
-        prompt += `  Line ${comment.lineIndex + 1}: ${comment.lineText}\n`;
-      } else if (
-        comment.lineText &&
-        comment.lineText.trim().length > 0 &&
-        !comment.isFileLevel
-      ) {
-        // Removed line or other content - show as removed; include original number when available
-        if (comment.isRemoved && comment.originalLineIndex !== undefined) {
-          prompt += `  Removed line ${comment.originalLineIndex}: ${comment.lineText}\n`;
-        } else {
-          prompt += `  Removed line: ${comment.lineText}\n`;
-        }
-      }
-      // For file headers (no meaningful lineText), just show the comment
-      prompt += `  Comment: ${comment.commentText}\n`;
-    });
-    prompt += "\n";
-  });
-
-  return prompt;
-}
 
 export default function DiffView({worktreePath, title = 'Diff Viewer', onClose, diffType = 'full', onAttachToSession, workspaceFeature}: Props) {
   const {rows: terminalHeight, columns: terminalWidth} = useTerminalDimensions();
